Move christmas intro screen to seasonal namespace
